Reorginize telemetry classes under the single namspace
