+++ lib/stoplight/domain/telemetry/light_registered.rb
@@ -2,7 +2,7 @@
 
 module Stoplight
   module Domain
-    class Telemetry
+    module Telemetry
       # Emitted once per (process instance, system, light) when the light's configuration first materializes.
       LightRegistered = Data.define(
         :settings
