+++ lib/stoplight/domain/telemetry/traffic_breached.rb
@@ -2,7 +2,7 @@
 
 module Stoplight
   module Domain
-    class Telemetry
+    module Telemetry
       # state_transitioned variant: traffic control tripped the light.
       TrafficBreached = Data.define(
         :from_color,
