+++ sig/stoplight/domain/telemetry/bus.rbs
@@ -1,0 +1,49 @@
+module Stoplight
+  module Domain
+    module Telemetry
+      # In-process telemetry bus.
+      #
+      # Contract:
+      # - Events are process-local observations: they describe what this process observed or did, not the
+      #   globally shared circuit state.
+      # - Every color change this process observes or effects emits exactly one +state_transitioned+ event. That
+      #   stream alone reconstructs the color timeline (breaches, recovery, and locks included).
+      # - Events are traffic-driven - no traffic, no events.
+      # - Handlers run synchronously on the calling thread and are error-isolated:
+      #   a raising handler is routed to the error notifier and never affects the run.
+      # - Emission allocates nothing when nothing is subscribed to that event class - a transitions-only
+      #   subscriber keeps RunCompleted emission free.
+      # - Publish pays for interested handlers only: no filter matching happens at publish time
+      #   (interest is resolved at subscribe time), and subscriptions to other event classes add
+      #   nothing. The handlers themselves run inline, so their cost belongs to the subscriber.
+      #   Subscribing mid-publish takes effect on the next publish.
+      class Bus
+        # The consumer side (subscribe/unsubscribe), what +_System#telemetry+ returns.
+        include _Telemetry
+
+        # The producer side (publish/subscribed?), what wiring injects into Emitter.
+        include _TelemetryPublisher
+
+        @error_notifier: ^(StandardError) -> void
+        @mutex: Mutex
+        @max_subscriptions: Integer
+        @subscriptions: Array[[Subscription, subscription_filter, subscription_handler]]
+        @dispatch: Hash[Class, Array[subscription_handler]]
+
+        # Default cap on live subscriptions.
+        MAX_SUBSCRIPTIONS: Integer
+
+        # @param error_notifier receives any StandardError a subscriber's handler raises.
+        # @param max_subscriptions caps the number of live subscriptions - #subscribe raises
+        #   Stoplight::Error::TooManySubscriptions once reached.
+        def initialize: (error_notifier: error_notifier, ?max_subscriptions: Integer) -> void
+
+        # Value-level mirror of the +event+ union. A new event type is added here and to the union
+        # together (a spec pins them in sync).
+        EVENT_CLASSES: Array[event_class]
+
+        private def rebuild: -> void
+      end
+    end
+  end
+end
