+++ sig/stoplight/domain/telemetry/recovery_failed.rbs
@@ -1,6 +1,6 @@
 module Stoplight
   module Domain
-    class Telemetry
+    module Telemetry
       # state_transitioned variant: recovery policy sent the light back to red.
       class RecoveryFailed < Data
         include StateTransitioned
