+++ sig/stoplight/domain/telemetry/recovery_started.rbs
@@ -1,6 +1,6 @@
 module Stoplight
   module Domain
-    class Telemetry
+    module Telemetry
       # state_transitioned variant: cool-off elapsed and the first probe explicitly entered recovery.
       class RecoveryStarted < Data
         include StateTransitioned
