+++ sig/stoplight/domain/telemetry/recovery_succeeded.rbs
@@ -1,6 +1,6 @@
 module Stoplight
   module Domain
-    class Telemetry
+    module Telemetry
       # state_transitioned variant: recovery policy resumed traffic.
       class RecoverySucceeded < Data
         include StateTransitioned
