+++ sig/stoplight/domain/telemetry/subscription.rbs
@@ -1,6 +1,6 @@
 module Stoplight
   module Domain
-    class Telemetry
+    module Telemetry
       # Opaque, single-use token returned by +Telemetry#subscribe+ and accepted by +Telemetry#unsubscribe+. No public behavior.
       class Subscription
       end
